Implement a quick fix to handle target elements that are null
This makes it possible to have the UIE version of the popup open while target elements are being destroyed (e.g. when changing a scene or if objects are deleted by script). However, the fix is not very elegant. When elements are removed, the window size should update, however, to calculate the required size, the popup content currently uses the GUI.skin.label property, which can only be accessed from within OnGUI. Ideally, this dependency is removed completely in the future which would also make handling resizing cleaner.